fix(auth): use BigDecimal for latitude and longitude to avoid floating point conversion issues
JPA was unable to map NUMERIC(10,8) columns to Double due to floating point
precision problems. Switching to BigDecimal preserves the exact decimal
representation in the database.

+++ src/main/java/tienda/uni/api/auth/persistence/entity/BuildingEntity.java
@@ -11,6 +11,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Getter
@@ -29,8 +30,8 @@ public class BuildingEntity {
     private String name;
 
     @Column(name = "latitude", nullable = false, precision = 10, scale = 8)
-    private Double latitude;
+    private BigDecimal latitude;
 
     @Column(name = "longitude", nullable = false, precision = 10, scale = 8)
-    private Double longitude;
+    private BigDecimal longitude;
 }
